refactor: rename shared helper module
