+++ tests/test-utils.test.ts
@@ -4,7 +4,7 @@ import path from 'node:path';
 
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import * as tg from '../src/utils/type-guards.js';
+import * as u from '../src/utils/shared-utils.js';
 import { cleanupTrackedTempDirs, useCleanMocks } from './test-utils.js';
 
 describe('cleanupTrackedTempDirs', () => {
@@ -15,8 +15,8 @@ describe('cleanupTrackedTempDirs', () => {
   });
 
   it('covers shared type guard helpers', () => {
-    expect([tg.isRecord({}), tg.ensureText(0)]).toEqual([true, '']);
-    expect(tg.normalizeBaseUrl(' https://x/// ')).toBe('https://x');
+    expect([u.isRecord({}), u.ensureText(0)]).toEqual([true, '']);
+    expect(u.normalizeBaseUrl(' https://x/// ')).toBe('https://x');
   });
 
   it('moves cwd out of a tracked directory before deleting it', () => {
